commenting to make file easier to look through
Created onClick handler function for converstaion LGItems, to seperate the function from the return portion of the comp.

also importing theme, to use in styled comp if needed.

Also trying to set up a preview of the most recent test for each converstaion, but need to figure out a better solution for truncating the text