added some comments to make file easier to look through at a glance.
Also had to rework the theme provider (hopefully one last time)
Didn't think about that calling a custom hook in multiple different places creates different instances of that state, which will not be shared across the whole app. So I had to rebuild the custom provider from before, but make a few minor tweaks to make sure it works properly.